XP integrada en mi propio usuario. Cuando la llamada a backend permita id, se hará con el resto de usuarios. (#108)

+++ app/src/main/java/com/gracehopper/laserchessapp/data/model/user/XPInfoResponse.kt
@@ -9,6 +9,6 @@ import com.google.gson.annotations.SerializedName
  * @property requiredXp Cantidad mínima de experiencia requerida para el siguiente nivel.
  */
 data class XPInfoResponse (
-    @SerializedName("xp") val xp: Long,
-    @SerializedName("required_xp") val requiredXp: Long
+    @SerializedName("xp") val xp: Int,
+    @SerializedName("required_xp") val requiredXp: Int
 )
